+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Unreleased
   extensions
 - Allowed entering of `base32` encoded strings containing spaces
 - Fixed pinentry dialog highlighting some messages incorrectly as errors
+- Fixed handling of empty user input through pinentry
 - Switched to using GitHub Actions as the project's CI pipeline
 - Updated minimum supported Rust version to `1.43.0`
 - Bumped `nitrokey` dependency to `0.9.0`